fix .net warnings :)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6550 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/LocalPlayer.cxx
@@ -1077,7 +1077,7 @@ void			LocalPlayer::jump()
   if (getFlag() == Flags::Wings)
 	newVelocity[2] = BZDB.eval(StateDatabase::BZDB_WINGSJUMPVELOCITY);
   else if (getFlag() == Flags::Bouncy)
-	newVelocity[2] = bzfrand() * BZDB.eval(StateDatabase::BZDB_JUMPVELOCITY);
+	newVelocity[2] = (float)(bzfrand() * BZDB.eval(StateDatabase::BZDB_JUMPVELOCITY));
   else
 	newVelocity[2] = BZDB.eval(StateDatabase::BZDB_JUMPVELOCITY);
   setVelocity(newVelocity);
